James: HTML updates and Css Matt: PHP and Sql Updates
James has adapted the html adding in a image as well as changing the colour of the tables as well as optimisation of the website. Matt has created a simple SQL database and has go the tables on the right hand side of the page to read from the database and display on the page by using a php query.