updated redirection
I have updated the redirection to be able to choose where to direct users each ad every time depending if they are logged in or not. I have also updated the name under the image of a logged in user to be able to be changed once a user has logged in. To personalize his experience.